Add beginning of LCD driver for the STM3240G-EVAL
git-svn-id: https://nuttx.svn.sourceforge.net/svnroot/nuttx/trunk@4627 7fd9a85b-ad96-42d3-883c-3090e2eb8679

+++ nuttx/configs/stm3240g-eval/src/up_selectsram.c
@@ -104,7 +104,7 @@ static const uint32_t g_sramconfig[] =
 
   GPIO_FSMC_NOE, GPIO_FSMC_NWE, GPIO_FSMC_NBL0, GPIO_FSMC_NBL1, GPIO_FSMC_NE2
 };
-#define NSRAM_CONFIG (sizeof(g_sramconfig)/sizeof(uint16_t))
+#define NSRAM_CONFIG (sizeof(g_sramconfig)/sizeof(uint32_t))
 
 /************************************************************************************
  * Private Functions
